Migrating onboarding folder to screens
More code formatting
