+++ mobile/lib/screens/auth/change_email_screen.dart
@@ -6,8 +6,8 @@ import 'package:app/widgets/text_fields.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
-import '../services/app_service.dart';
-import '../services/firebase_service.dart';
+import '../../services/app_service.dart';
+import '../../services/firebase_service.dart';
 
 class ChangeEmailScreen extends StatefulWidget {
   const ChangeEmailScreen({
