+++ mobile/lib/screens/auth/phone_reauthenticate_screen.dart
@@ -7,7 +7,7 @@ import 'package:app/widgets/text_fields.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
-import '../utils/network.dart';
+import '../../utils/network.dart';
 
 class PhoneReAuthenticateScreen extends StatefulWidget {
   final UserDetails userDetails;
