+++ mobile/lib/screens/favourite_places/favourite_places_page.dart
@@ -1,12 +1,12 @@
 import 'package:app/constants/config.dart';
 import 'package:app/models/place_details.dart';
-import 'package:app/screens/search/search_page.dart';
 import 'package:app/services/app_service.dart';
 import 'package:app/widgets/custom_widgets.dart';
 import 'package:app/widgets/favourite_place_card.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/svg.dart';
 import 'package:provider/provider.dart';
+
+import 'favourite_places_widgets.dart';
 
 class FavouritePlaces extends StatefulWidget {
   const FavouritePlaces({Key? key}) : super(key: key);
@@ -27,7 +27,7 @@ class _FavouritePlacesState extends State<FavouritePlaces> {
           child: Consumer<PlaceDetailsModel>(
             builder: (context, placeDetailsModel, child) {
               if (placeDetailsModel.favouritePlaces.isEmpty) {
-                return _emptyPlaces();
+                return const EmptyFavouritePlaces();
               }
 
               return refreshIndicator(
@@ -42,55 +42,6 @@ class _FavouritePlacesState extends State<FavouritePlaces> {
                   onRefresh: _refreshPage);
             },
           )),
-    );
-  }
-
-  Widget _emptyPlaces() {
-    return Container(
-      color: Config.appBodyColor,
-      padding: const EdgeInsets.all(40.0),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          RichText(
-              textAlign: TextAlign.center,
-              text: TextSpan(children: [
-                TextSpan(
-                    text: 'Tap the ',
-                    style: Theme.of(context).textTheme.bodyText1),
-                WidgetSpan(
-                    child: SvgPicture.asset(
-                  'assets/icon/heart.svg',
-                  semanticsLabel: 'Favorite',
-                  height: 15.33,
-                  width: 15.12,
-                )),
-                TextSpan(
-                    text: ' Favorite icon on any location air quality '
-                        'to save them here for later.',
-                    style: Theme.of(context).textTheme.bodyText1),
-              ])),
-          const SizedBox(
-            height: 10,
-          ),
-          OutlinedButton(
-            onPressed: () async {
-              await Navigator.push(context,
-                  MaterialPageRoute(builder: (context) {
-                return const SearchPage();
-              }));
-            },
-            style: OutlinedButton.styleFrom(
-              shape: const CircleBorder(),
-              padding: const EdgeInsets.all(24),
-            ),
-            child: Text(
-              'Add',
-              style: TextStyle(color: Config.appColor),
-            ),
-          )
-        ],
-      ),
     );
   }
