+++ mobile/lib/screens/kya/kya_widgets.dart
@@ -1,9 +1,12 @@
 import 'package:auto_size_text/auto_size_text.dart';
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
 
 import '../../constants/config.dart';
 import '../../models/kya.dart';
 import '../../themes/light_theme.dart';
+import 'kya_title_page.dart';
 
 String getKyaMessage({required Kya kya}) {
   var kyaItems = kya.lessons.length;
@@ -64,6 +67,91 @@ class KyaProgressBar extends StatelessWidget {
             color: Config.appColorBlue,
             value: kya.progress / kya.lessons.length,
             backgroundColor: Config.appColorDisabled.withOpacity(0.2),
+          )),
+    );
+  }
+}
+
+class KyaViewWidget extends StatelessWidget {
+  final Kya kya;
+  const KyaViewWidget({Key? key, required this.kya}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(top: 8),
+      child: GestureDetector(
+          onTap: () async {
+            await Navigator.push(context, MaterialPageRoute(builder: (context) {
+              return KyaTitlePage(kya);
+            }));
+          },
+          child: Container(
+            padding: const EdgeInsets.fromLTRB(16.0, 8.0, 8.0, 8.0),
+            decoration: const BoxDecoration(
+                color: Colors.white,
+                borderRadius: BorderRadius.all(Radius.circular(16.0))),
+            child: Row(
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: [
+                Expanded(
+                  child: Column(
+                    children: [
+                      AutoSizeText(
+                        kya.title,
+                        maxLines: 4,
+                        overflow: TextOverflow.ellipsis,
+                        style: CustomTextStyle.headline10(context),
+                      ),
+                      const SizedBox(
+                        height: 28,
+                      ),
+                      Row(
+                        crossAxisAlignment: CrossAxisAlignment.center,
+                        mainAxisAlignment: MainAxisAlignment.start,
+                        children: [
+                          KyaMessage(
+                            kya: kya,
+                          ),
+                          const SizedBox(
+                            width: 6,
+                          ),
+                          SvgPicture.asset(
+                            'assets/icon/more_arrow.svg',
+                            semanticsLabel: 'more',
+                            height: 6.99,
+                            width: 4,
+                          ),
+                        ],
+                      ),
+                      SizedBox(
+                        height:
+                            getKyaMessage(kya: kya).toLowerCase() == 'continue'
+                                ? 2
+                                : 0,
+                      ),
+                      KyaProgressBar(kya: kya),
+                    ],
+                  ),
+                ),
+                const SizedBox(
+                  width: 16,
+                ),
+                Container(
+                  width: 104,
+                  height: 104,
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(8.0),
+                    image: DecorationImage(
+                      fit: BoxFit.cover,
+                      image: CachedNetworkImageProvider(
+                        kya.imageUrl,
+                      ),
+                    ),
+                  ),
+                ),
+              ],
+            ),
           )),
     );
   }
