+++ mobile/lib/screens/on_boarding/location_setup_screen.dart
@@ -1,14 +1,14 @@
 import 'package:app/constants/config.dart';
 import 'package:app/models/enum_constants.dart';
-import 'package:app/on_boarding/setup_complete_screeen.dart';
 import 'package:app/screens/home_page.dart';
+import 'package:app/screens/on_boarding/setup_complete_screeen.dart';
 import 'package:app/utils/dialogs.dart';
 import 'package:app/widgets/buttons.dart';
 import 'package:flutter/material.dart';
 
-import '../services/local_storage.dart';
-import '../services/native_api.dart';
-import '../themes/light_theme.dart';
+import '../../services/local_storage.dart';
+import '../../services/native_api.dart';
+import '../../themes/light_theme.dart';
 
 class LocationSetupScreen extends StatefulWidget {
   const LocationSetupScreen({Key? key}) : super(key: key);
