+++ mobile/lib/screens/on_boarding/profile_setup_screen.dart
@@ -10,9 +10,9 @@ import 'package:app/widgets/text_fields.dart';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 
-import '../models/enum_constants.dart';
-import '../services/local_storage.dart';
-import '../themes/light_theme.dart';
+import '../../models/enum_constants.dart';
+import '../../services/local_storage.dart';
+import '../../themes/light_theme.dart';
 import 'notifications_setup_screen.dart';
 
 class ProfileSetupScreen extends StatefulWidget {
