+++ mobile/lib/screens/on_boarding/spash_screen.dart
@@ -1,18 +1,18 @@
 import 'package:animations/animations.dart';
-import 'package:app/auth/phone_auth_widget.dart';
-import 'package:app/on_boarding/location_setup_screen.dart';
-import 'package:app/on_boarding/notifications_setup_screen.dart';
-import 'package:app/on_boarding/profile_setup_screen.dart';
-import 'package:app/on_boarding/setup_complete_screeen.dart';
-import 'package:app/on_boarding/welcome_screen.dart';
 import 'package:app/screens/home_page.dart';
+import 'package:app/screens/on_boarding/profile_setup_screen.dart';
+import 'package:app/screens/on_boarding/setup_complete_screeen.dart';
+import 'package:app/screens/on_boarding/welcome_screen.dart';
 import 'package:app/services/app_service.dart';
 import 'package:app/utils/extensions.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 
-import '../models/enum_constants.dart';
-import '../services/local_storage.dart';
+import '../../models/enum_constants.dart';
+import '../../services/local_storage.dart';
+import '../auth/phone_auth_widget.dart';
+import 'location_setup_screen.dart';
+import 'notifications_setup_screen.dart';
 
 class SplashScreen extends StatefulWidget {
   const SplashScreen({Key? key}) : super(key: key);
