+++ mobile/pubspec.yaml
@@ -69,6 +69,7 @@ dependencies:
   webview_flutter: ^3.0.2
   hive: ^2.1.0
   hive_flutter: ^1.1.0
+  firebase_in_app_messaging: ^0.6.0+14
   # https://pub.dev/packages/permission_handler
   # https://pub.dev/packages/flutter_cache_manager
   # https://pub.dev/packages/http_retry
